app/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
pdfminer.six==20201018
pandas==1.1.3
numpy==1.18.5
spacy==2.3.4
tensorflow-cpu==2.3.2
transformers==4.3.2
